+++ app/src/main/java/com/example/bluettoothmatching/bluetooth/BlutoothBK.kt
@@ -58,8 +58,8 @@ class BlutoothBK : Service() {
 
                     MacAddressList.clear()
                     MacAddressList.addAll(MacAddressSet)
-                    MacAddressList.add("80:9F:F5:79:8C:E2")
-                    MacAddressList.add("58:C6:F0:5C:85:68")
+                    //MacAddressList.add("80:9F:F5:79:8C:E2")
+                    //MacAddressList.add("58:C6:F0:5C:85:68")
                     //MacAddressList.add("1")
                     //MacAddressList.add("2")
                     //MacAddressList.add("3")
