+++ app/src/main/java/com/example/bluettoothmatching/database/Profile.kt
@@ -4,5 +4,5 @@ data class Profile(
     var address: String? = null,
     var name: String? = null,
     var message: String? = null,
-    //var bitmap: String? = null
+    var bitmap: String? = null
 )
